generate-password/requirements.txt
cryptography
qrcode
psycopg2-binary
Pillow